Moved all logic into most basic class
